requirements.txt
streamlit==1.29.0
pandas==2.1.3
plotly==5.18.0
gspread==5.12.0
google-auth==2.23.4
python-dotenv==1.0.0